Refactored code by consolidating duplicate name and desc methods into a BaseMetadata class and updated Metadata and Track to extend it.
